TOMLConfig.jl

API Reference

TOMLConfig

ArgParse.parse_args

TOMLConfig.parse_config

TOMLConfig.save_config

ArgParse.add_arg_table!

TOMLConfig.parser_settings!

TOMLConfig.Config

TOMLConfig.ArgTableEntry

TOMLConfig.defaults!

TOMLConfig.arg_table_flag

TOMLConfig — Module

TOMLConfig

Use TOML files to configure command line parsing via ArgParse.jl.

ArgParse.parse_args — Function

ArgParse.parse_args(
    [args_list::Vector = ARGS,]
    [settings::ArgParseSettings = ArgParseSettings(),]
    cfg::Config;
    as_dict = false,
    as_symbols = false,
    kwargs...
)

Parse TOML configuration struct with command line arguments args_list.

Arguments:

args_list::Vector: (optional) vector of arguments to be parsed

settings::ArgParseSettings: (optional) settings struct which will be configured according to cfg

cfg::Config: TOML configuration settings

Keywords:

as_dict: if true, return config as a dictionary with String keys, otherwise return a Config struct

as_symbols: if true and as_dict=true, return config dictionary with Symbol keys

kwargs...: additional keyword arguments can be passed to customize the global parser settings; see TOMLConfig.parser_settings!. Note that these settings remain persistent on subsequent calls to parse_args; use TOMLConfig.parser_settings!() to restore all defaults.

Examples

julia> using TOMLConfig

julia> cfg = Config(TOML.parse(
       """
       a = 1
       b = 2
       [sec1]
       b = "_PARENT_"
       c = 3
           [sec1.sub1]
           _INHERIT_ = "_PARENT_"
       """));

julia> parsed_args = parse_args(["--a", "3", "--sec1.b", "5", "--sec1.c", "10"], cfg)
TOML Config with contents:

b = 2
a = 3

[sec1]
c = 10
b = 5

    [sec1.sub1]
    c = 10
    b = 5

TOMLConfig.parse_config — Function

parse_config(args...; filename::String, kwargs...)

Convenience method for parsing configuration files. Equivalent to ArgParse.parse_args(args..., TOML.parsefile(Config, filename); kwargs...).

TOMLConfig.save_config — Function

save_config(cfg::Union{Config, AbstractDict}; filename::AbstractString, kwargs...)

Save configuration cfg to TOML file filename. Additional keyword arguments are forwarded to TOML.print.

ArgParse.add_arg_table! — Function

ArgParse.add_arg_table!(settings::ArgParseSettings, cfg::Config)

Populate settings argument table using configuration cfg.

Examples

julia> using TOMLConfig

julia> cfg = Config(TOML.parse(
       """
       a = 1.0
       b = 2
       [sec1]
       c = [3, 4]
           [sec1.sub1]
           d = "d"
       """));

julia> settings = add_arg_table!(ArgParseSettings(prog = "myscript.jl"), cfg);

julia> ArgParse.show_help(settings; exit_when_done = false)
usage: myscript.jl [--b B] [--a A] [--sec1.c [SEC1.C...]]
                   [--sec1.sub1.d SEC1.SUB1.D]

optional arguments:
  --b B                 (type: Int64, default: 2)
  --a A                 (type: Float64, default: 1.0)
  --sec1.c [SEC1.C...]  (type: Int64, default: [3, 4])
  --sec1.sub1.d SEC1.SUB1.D
                        (default: "d")

TOMLConfig.parser_settings! — Function

parser_settings!(;
    arg_key = "_ARG_",
    arg_required_value = "_REQUIRED_",
    flag_delim = ".",
    inherit_all_key = "_INHERIT_",
    inherit_parent_value = "_PARENT_",
)

Customize global settings for parsing Config structs. To restore default settings, use TOMLConfig.parser_settings!().

Keywords:

arg_key: if this key is found in a TOML section, the rest of the section is interpreted as properties for the argument table entry

arg_required_value: if this value is found in a TOML section, the required = true is passed to the argument table entry

inherit_all_key: if this key is found in a TOML section, all fields from the immediate parent section (i.e., non-recursive) should be inherited

inherit_parent_value: if this value is found in a TOML section, it is replaced with the value corresponding to the same key in the immediate parent section (i.e., non-recursive)

flag_delim: command line flags for keys in nested TOML sections are formed by joining all parent keys together with this delimiter

TOMLConfig.Config — Type

Config(toml::AbstractDict{String})

Basic tree structure for navigating TOML file contents. Each node in the Config tree represents a single section of a TOML file. Children of a Config node are the corresponding TOML subsections, if they exist.

Examples

Construct a Config from a TOML string literal, parsed using the TOML standard library:

julia> using TOMLConfig

julia> cfg = Config(TOML.parse(
       """
       a = 1
       [sec1]
           b = 2
           [sec1.sub1]
           c = 3
       """))
TOML Config with contents:

a = 1

[sec1]
b = 2

    [sec1.sub1]
    c = 3

julia> cfg.sec1
TOML Config with contents:

b = 2

[sub1]
c = 3

julia> cfg.sec1.sub1
TOML Config with contents:

c = 3

Config can also be constructed incrementally:

julia> using TOMLConfig

julia> cfg = Config();

julia> cfg.a = 1;

julia> cfg.b.c.d = "three"; # Nested fields will be dynamically set, if they don't exist

julia> cfg["b"]["e"] = [2]; # getindex!/setindex! and getproperty!/setproperty! are all supported

julia> cfg
TOML Config with contents:

a = 1

[b]
e = [2]

    [b.c]
    d = "three"

TOMLConfig.ArgTableEntry — Type

ArgTableEntry

Internal struct used to define leaf values which represent ArgParse argument table entries, storing e.g. properties of the corresponding command line argument; see ArgParse.add_arg_table!.

TOMLConfig.defaults! — Function

defaults!(cfg::Config)

Populate fields of TOML config which are specified to have default values inherited from parent sections.

Examples

julia> using TOMLConfig

julia> cfg = TOMLConfig.defaults!(Config(TOML.parse(
       """
       a = 1
       b = 2
       [sec1]
       b = "_PARENT_"
       c = 3
           [sec1.sub1]
           _INHERIT_ = "_PARENT_"
       """)))
TOML Config with contents:

b = 2
a = 1

[sec1]
c = 3
b = 2

    [sec1.sub1]
    c = 3
    b = 2

TOMLConfig.arg_table_flag — Function

arg_table_flag(cfg::Config, k::String)

Generate command flag corresponding to nested key k in a Config node. The flag is constructed by joining the keys recursively from the parents of the current node using the delimiter flag_delim() and prepending "–".

Examples

julia> using TOMLConfig

julia> cfg = Config();

julia> cfg.a = 1;

julia> cfg.sec1.b = 2;

julia> cfg.sec1.sub1.c = 3;

julia> cfg
TOML Config with contents:

a = 1

[sec1]
b = 2

    [sec1.sub1]
    c = 3

julia> TOMLConfig.arg_table_flag(cfg, "a")
"--a"

julia> TOMLConfig.arg_table_flag(cfg.sec1, "b")
"--sec1.b"

julia> TOMLConfig.arg_table_flag(cfg.sec1.sub1, "c")
"--sec1.sub1.c"